+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [5.1.0] - 2021-04-21
+- Updated to work in MySQL 8.0 environment.
 - Validated to work in MySQL 5.7 environment.
 
 ### Changed
+- mysql_class.Server.upd_srv_stat:  Set qry_cache attribute accordingly to the version setting.
 - mysql_class.Server.connect:  Added get_server_version call to set server's version attribute.
 - mysql_class.Server.\_\_init\_\_:  Added version attribute to hold MySQL server's version.