+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated to work in MySQL 8.0 environment.
 - Validated to work in MySQL 5.7 environment.
 
+### Fixed
+- mysql_class.SlaveRep.\_\_init\_\_:  Renamed conflicted self.retry to self.conn_retry and self.tran_retry.
+
 ### Changed
 - mysql_class.Server.upd_mst_rep_stat:  Set innodb_xa to None if not present in MySQL 8.0.
 - mysql_class.Server.upd_srv_stat:  Set qry_cache to zero if not present in MySQL 8.0.