+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated to work in MySQL 5.7 environment.
 
 ### Fixed
+- mysql_class.Server.connect:  Set self.conn_msg to None if login is successful.
 - mysql_class.SlaveRep.get_others:  Renamed self.retry to self.tran_retry.
 - mysql_class.SlaveRep.upd_slv_status:  Renamed conflicted self.retry to self.conn_retry and self.tran_retry.
 - mysql_class.SlaveRep.\_\_init\_\_:  Renamed conflicted self.retry to self.conn_retry and self.tran_retry.
@@ -17,7 +18,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - mysql_class.Server.set_pass_config:  Set the SQL passwd config attributes.
 
 ### Changed
-- mysql_class.Server.\_\_init\_\_:  Added call to set_pass_config and removed setting of sql_pass and config attributes.
+- mysql_class.Server.\_\_init\_\_:  Added call to set_pass_config and removed setting of self.sql_pass and self.config attributes.
 - mysql_class.SlaveRep.upd_slv_status:  The attributes tran_retry and run are in different location in MySQL 8.0.
 - mysql_class.Server.upd_mst_rep_stat:  Set innodb_xa to None if not present in MySQL 8.0.
 - mysql_class.Server.upd_srv_stat:  Set qry_cache to zero if not present in MySQL 8.0.